scripts/youtube-countertrade/requirements.txt
youtube-transcript-api>=1.2.0
requests>=2.31.0
feedparser>=6.0.0
python-dotenv>=1.0.0
